+++ sphere.h
@@ -20,7 +20,7 @@ __device__ bool sphere::hit(const ray& r, float t_min, float t_max, hit_record& 
     float c = dot(oc, oc) - radius*radius;
     float discriminant = b*b - a*c;
     if (discriminant > 0) {
-        float temp = (-b - sqrt(discriminant))/a;
+        float temp = (-b - sqrtf(discriminant))/a;
         if (temp < t_max && temp > t_min) {
             rec.t = temp;
             rec.p = r.point_at_parameter(rec.t);
@@ -28,7 +28,7 @@ __device__ bool sphere::hit(const ray& r, float t_min, float t_max, hit_record& 
             rec.mat_ptr = mat_ptr;
             return true;
         }
-        temp = (-b + sqrt(discriminant)) / a;
+        temp = (-b + sqrtf(discriminant)) / a;
         if (temp < t_max && temp > t_min) {
             rec.t = temp;
             rec.p = r.point_at_parameter(rec.t);
